Tooling types now moved to primitive type mapping.
